+++ contracts/VRFv2SubscriptionConsumer.sol
@@ -56,7 +56,7 @@ abstract contract VRFv2SubscriptionConsumer is VRFConsumerBaseV2, ConfirmedOwner
     // Assumes the subscription is funded sufficiently.
     function requestRandomWords(uint32 numWords) internal returns (uint)    {
         // Will revert if subscription is not set and funded.
-        uint requestId = COORDINATOR.requestRandomWords(
+        uint256 requestId = COORDINATOR.requestRandomWords(
             hashKey,
             subscriptionId,
             requestConfirmations,
